refactor: split modules into packages with isolated responsibilities
