+++ huoguoml/server/api/ml_model.py
@@ -1,4 +1,4 @@
-from typing import List, Optional
+from typing import List, Optional, Any
 
 import requests
 from fastapi import APIRouter, HTTPException, Request, Query
@@ -7,17 +7,14 @@ from starlette.responses import FileResponse
 
 from huoguoml.schema.ml_model import MLModelIn, MLModel, MLModelRegistry, MLModelTag
 from huoguoml.schema.ml_service import MLService
-from huoguoml.server.db.service import Service
+from huoguoml.server.service.ml_model import MLModelService
 from huoguoml.util.string import concat_uri, coerce_url
 
 
-class MLModelRouter(object):
+class MLModelRouter(APIRouter):
 
-    def __init__(self, service: Service):
-        router = APIRouter(
-            prefix="/api/models",
-            tags=["models"],
-        )
+    def __init__(self, service: MLModelService, **extra: Any):
+        super().__init__(**extra, prefix="/api/models", tags=["models"])
 
         def notify_services(ml_model_name: str, ml_model_rule: str):
             ml_services = service.update_ml_services_by_model_name(ml_model_name=ml_model_name,
@@ -28,11 +25,11 @@ class MLModelRouter(object):
                 update_api = concat_uri(uri, "api", "update")
                 requests.post(update_api, json=ml_service_obj.dict()).raise_for_status()
 
-        @router.get("", response_model=List[MLModelRegistry])
+        @self.get("", response_model=List[MLModelRegistry])
         async def get_ml_models_groupby_name():
             return service.get_ml_models_groupby_name()
 
-        @router.post("", response_model=MLModel)
+        @self.post("", response_model=MLModel)
         async def create_ml_model(ml_model_in: MLModelIn, background_tasks: BackgroundTasks):
             ml_model = service.create_ml_model(ml_model_in=ml_model_in)
             background_tasks.add_task(notify_services, ml_model_name=ml_model_in.name, ml_model_rule="latest")
@@ -40,7 +37,7 @@ class MLModelRouter(object):
                 raise HTTPException(status_code=400)
             return ml_model
 
-        @router.put("/{ml_model_name}/{ml_model_version}", response_model=MLModel)
+        @self.put("/{ml_model_name}/{ml_model_version}", response_model=MLModel)
         async def update_ml_model(ml_model_name: str, ml_model_version: str, ml_model: MLModel,
                                   background_tasks: BackgroundTasks):
             ml_model = service.update_ml_model(ml_model_name=ml_model_name,
@@ -58,7 +55,7 @@ class MLModelRouter(object):
                 raise HTTPException(status_code=400)
             return ml_model
 
-        @router.get("/{ml_model_name}/{ml_model_version}", response_model=MLModel, responses={
+        @self.get("/{ml_model_name}/{ml_model_version}", response_model=MLModel, responses={
             200: {
                 "content": {"application/zip": {}},
                 "description": "Return the model as json or models files as zip",
@@ -76,12 +73,12 @@ class MLModelRouter(object):
                 response = FileResponse(file_path, media_type='application/zip')
             return response
 
-        @router.get("/{ml_model_name}", response_model=MLModel)
-        async def get_ml_model(ml_model_name: str,
-                               tag: Optional[str] = Query(None,
-                                                          description="Model tag as string (available: "
-                                                                      "production, staging"),
-                               ):
+        @self.get("/{ml_model_name}", response_model=List[MLModel])
+        async def get_ml_models(ml_model_name: str,
+                                tag: Optional[str] = Query(None,
+                                                           description="Model tag as string (available: "
+                                                                       "production, staging"),
+                                ):
             if tag is not None:
                 ml_model = service.get_ml_model_by_tag(ml_model_name=ml_model_name,
                                                        tag=tag)
@@ -91,11 +88,8 @@ class MLModelRouter(object):
                                                                            ml_model_version=ml_model.version)
 
                 return FileResponse(file_path, media_type='application/zip')
-
             else:
                 ml_models = service.get_ml_models_by_name(ml_model_name=ml_model_name)
                 if not ml_models:
                     raise HTTPException(status_code=404)
                 return ml_models
-
-        self.router = router
