+++ huoguoml/server/api/ml_service.py
@@ -1,19 +1,19 @@
-from typing import List
+from typing import List, Any
 
-from fastapi import APIRouter, HTTPException, Request
+from fastapi import HTTPException, Request, APIRouter
 
 from huoguoml.schema.ml_service import MLService, MLServiceIn
-from huoguoml.server.db.service import Service
+from huoguoml.server.service.ml_service import MLServiceService
 
 
-class MLServiceRouter(object):
-    def __init__(self, service: Service):
-        router = APIRouter(
-            prefix="/api/services",
-            tags=["services"],
-        )
+class MLServiceRouter(APIRouter):
+    def __init__(self, service: MLServiceService, **extra: Any):
+        super().__init__(**extra,
+                         prefix="/api/services",
+                         tags=["services"],
+                         )
 
-        @router.put("", response_model=MLService)
+        @self.put("", response_model=MLService)
         async def create_or_update_ml_service(ml_service_in: MLServiceIn, request: Request):
             ml_service_in.host = request.client.host
             ml_service = service.create_or_update_ml_service(ml_service_in=ml_service_in)
@@ -21,15 +21,13 @@ class MLServiceRouter(object):
                 raise HTTPException(status_code=400)
             return ml_service
 
-        @router.get("", response_model=List[MLService])
+        @self.get("", response_model=List[MLService])
         async def get_ml_services():
             return service.get_ml_services()
 
-        @router.get("/{service_id}", response_model=MLService)
+        @self.get("/{service_id}", response_model=MLService)
         async def get_ml_service(service_id: int):
             ml_service = service.get_ml_service(service_id=service_id)
             if not ml_service:
                 raise HTTPException(status_code=404)
             return ml_service
-
-        self.router = router
