+++ huoguoml/server/entity/__init__.py
@@ -1,0 +1,6 @@
+"""
+The huoguoml.entity module contains the model definition for our ORM mapper SQL Alchemy
+"""
+from sqlalchemy.orm import declarative_base
+
+Base = declarative_base()
